Move ExpandableText into separate component
